6 семестр/Основы промышленной разработки программного обеспечения/server/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic[email]~=2.6.4
motor==3.3.2
python-dotenv~=1.0.1
starlette~=0.36.3
pymongo~=4.6.2
pytest~=8.2.1
pytest-asyncio~=0.23.7